Mñeh mñeh

+++ Cell.java
@@ -26,8 +26,9 @@ public class Cell implements Comparable<Cell> {
 		float c_n=1.0f;//non diagonal
 		float c_d=1.414f; //diagonal
 		float h= (c_d*d_min) + (c_n*(d_max-d_min));
-		return h;
+		return (float) h;
 		*/
+		//return (float) Math.abs(this.x - other.x) + Math.abs(this.y - other.y);
 		return (float) (Math.pow(this.x - other.x, 2) + Math.pow(this.y - other.y, 2));
 	}
